+++ crates/orbit-store/src/file/task_store/v2/sidecars.rs
@@ -1,0 +1,77 @@
+use super::*;
+
+impl TaskV2Store {
+    pub(crate) fn get_task_comments(
+        &self,
+        id: &str,
+    ) -> Result<Option<Vec<TaskComment>>, OrbitError> {
+        orbit_common::types::validate_orb_task_id(id)?;
+        match self.bundle_store.read_bundle(id) {
+            Ok(bundle) => Ok(Some(
+                bundle
+                    .comments
+                    .into_iter()
+                    .map(|comment| TaskComment {
+                        at: comment.at,
+                        by: comment.by,
+                        message: comment.body,
+                    })
+                    .collect(),
+            )),
+            Err(OrbitError::NotFound {
+                kind: NotFoundKind::Task,
+                ..
+            }) => Ok(None),
+            Err(err) => Err(err),
+        }
+    }
+
+    pub(crate) fn get_task_history(
+        &self,
+        id: &str,
+    ) -> Result<Option<Vec<TaskHistoryEntry>>, OrbitError> {
+        orbit_common::types::validate_orb_task_id(id)?;
+        match self.bundle_store.read_bundle(id) {
+            Ok(bundle) => Ok(Some(
+                bundle
+                    .events
+                    .into_iter()
+                    .map(|event| TaskHistoryEntry {
+                        at: event.at,
+                        by: event.by,
+                        event: event.event_type,
+                        note: event.note,
+                        from_status: event.from_status,
+                        to_status: event.to_status,
+                    })
+                    .collect(),
+            )),
+            Err(OrbitError::NotFound {
+                kind: NotFoundKind::Task,
+                ..
+            }) => Ok(None),
+            Err(err) => Err(err),
+        }
+    }
+
+    pub(crate) fn get_task_review_threads(
+        &self,
+        id: &str,
+    ) -> Result<Option<Vec<ReviewThread>>, OrbitError> {
+        orbit_common::types::validate_orb_task_id(id)?;
+        match self.bundle_store.read_bundle(id) {
+            Ok(bundle) => Ok(Some(
+                bundle
+                    .review_threads
+                    .into_iter()
+                    .map(review_thread_from_v2)
+                    .collect(),
+            )),
+            Err(OrbitError::NotFound {
+                kind: NotFoundKind::Task,
+                ..
+            }) => Ok(None),
+            Err(err) => Err(err),
+        }
+    }
+}
