+++ crates/orbit-store/src/file/task_store/v2/tests/mod.rs
@@ -1,0 +1,80 @@
+use chrono::{TimeZone, Utc};
+use orbit_common::types::{
+    ExternalRef, ReviewMessage, ReviewThread, ReviewThreadStatus, TaskArtifact, TaskComment,
+    TaskHistoryEntry, TaskPriority, TaskStatus, TaskType,
+};
+use tempfile::TempDir;
+
+use super::*;
+use crate::backend::{
+    TaskArtifactUpdateParams, TaskCreateParams, TaskDocumentUpdateParams, TaskHistoryUpdateParams,
+    TaskReviewUpdateParams,
+};
+use crate::sqlite::task_registry::{BindWorkspaceParams, TaskRegistryStore, task_registry_path};
+
+pub(super) fn store(temp: &TempDir) -> TaskV2Store {
+    let registry =
+        TaskRegistryStore::open(&task_registry_path(temp.path())).expect("open registry");
+    let repo_dir = temp.path().join("repo");
+    let orbit_dir = repo_dir.join(".orbit");
+    std::fs::create_dir_all(&orbit_dir).expect("create orbit dir");
+    let binding = registry
+        .bind_workspace(BindWorkspaceParams {
+            workspace_id: Some("orbit-test-123456".to_string()),
+            slug: "Orbit Test".to_string(),
+            repo_root: repo_dir.clone(),
+            workspace_path: repo_dir.clone(),
+            orbit_dir: orbit_dir.clone(),
+            repo_fingerprint: None,
+        })
+        .expect("bind workspace");
+    TaskV2Store::new(
+        registry,
+        binding.workspace_id,
+        orbit_dir,
+        Some(repo_dir.to_string_lossy().into_owned()),
+        Some(repo_dir.to_string_lossy().into_owned()),
+    )
+}
+
+pub(super) fn create_params(title: &str, status: TaskStatus) -> TaskCreateParams {
+    let now = Utc.with_ymd_and_hms(2026, 5, 11, 12, 0, 0).unwrap();
+    TaskCreateParams {
+        actor: "codex:gpt-5.5".to_string(),
+        parent_id: None,
+        title: title.to_string(),
+        description: "Detailed task description".to_string(),
+        acceptance_criteria: vec![
+            "First criterion".to_string(),
+            "Second criterion".to_string(),
+        ],
+        dependencies: Vec::new(),
+        relations: Vec::new(),
+        tags: vec!["task-artifacts".to_string(), "v2".to_string()],
+        plan: "1. Do the work".to_string(),
+        execution_summary: String::new(),
+        context_files: vec!["docs/design/task-artifacts/_plan.md".to_string()],
+        workspace_path: None,
+        repo_root: None,
+        created_by: Some("codex:gpt-5.5".to_string()),
+        planned_by: None,
+        implemented_by: None,
+        status,
+        priority: TaskPriority::High,
+        complexity: None,
+        task_type: TaskType::Feature,
+        external_refs: vec![
+            ExternalRef::try_new("linear".to_string(), "ENG-123".to_string(), None).unwrap(),
+        ],
+        source_task_id: None,
+        crew: None,
+        comments: vec![TaskComment {
+            at: now,
+            by: "daniel".to_string(),
+            message: "Please build this.".to_string(),
+        }],
+    }
+}
+
+mod crud_tests;
+mod update_tests;
